final push in the pg-move
