+++ martin/src/config/args/postgres.rs
@@ -10,7 +10,7 @@ use serde::{Deserialize, Serialize};
 use super::connections::Arguments;
 use super::connections::State::{Ignore, Take};
 use crate::config::file::UnrecognizedValues;
-use crate::config::file::pg::{POOL_SIZE_DEFAULT, PgConfig, PgSslCerts};
+use crate::config::file::postgres::{POOL_SIZE_DEFAULT, PgConfig, PgSslCerts};
 // Must match the help string for BoundsType::Quick
 pub const DEFAULT_BOUNDS_TIMEOUT: Duration = Duration::from_secs(5);
