+++ martin/src/config/file/mod.rs
@@ -12,10 +12,10 @@ pub mod cog;
 pub mod fonts;
 #[cfg(feature = "mbtiles")]
 pub mod mbtiles;
-#[cfg(feature = "postgres")]
-pub mod pg;
 #[cfg(feature = "pmtiles")]
 pub mod pmtiles;
+#[cfg(feature = "postgres")]
+pub mod postgres;
 #[cfg(feature = "sprites")]
 pub mod sprites;
 #[cfg(feature = "styles")]
