+++ martin/src/config/file/postgres/config_function.rs
@@ -5,7 +5,7 @@ use tilejson::{Bounds, TileJSON};
 
 use super::config::PgInfo;
 use crate::config::file::UnrecognizedValues;
-use crate::config::file::pg::utils::patch_json;
+use crate::config::file::postgres::utils::patch_json;
 
 pub type FuncInfoSources = BTreeMap<String, FunctionInfo>;
