+++ martin/src/config/file/postgres/config_table.rs
@@ -6,8 +6,7 @@ use tilejson::{Bounds, TileJSON, VectorLayer};
 
 use super::PgInfo;
 use crate::config::file::UnrecognizedValues;
-use crate::config::file::pg::utils::patch_json;
-use crate::pg::utils::normalize_key;
+use crate::config::file::postgres::utils::{normalize_key, patch_json};
 
 pub type TableInfoSources = BTreeMap<String, TableInfo>;
