+++ martin/src/config/file/postgres/mod.rs
@@ -7,4 +7,9 @@ pub use config_function::*;
 mod config_table;
 pub use config_table::*;
 
-mod utils;
+pub(crate) mod utils;
+
+mod builder;
+pub use builder::PgBuilder;
+
+pub mod resolver;
