+++ martin/src/config/file/postgres/resolver/mod.rs
@@ -1,0 +1,5 @@
+mod query_functions;
+mod query_tables;
+
+pub use query_functions::query_available_function;
+pub use query_tables::{query_available_tables, table_to_query};
