+++ martin/src/config/file/postgres/resolver/scripts/query_available_tables.sql
