--- martin/src/pg/mod.rs
@@ -1,7 +1,0 @@
-pub(crate) mod builder;
-mod query_functions;
-mod query_tables;
-pub(crate) mod utils;
-
-pub use query_functions::query_available_function;
-pub use query_tables::query_available_tables;
